Version 1.96
Make sure the special characters "&", "<", ">", "|", and "^" are handled correctly if shell is cmd.exe and shellxquote is \" (there will be miscellaneous gotchas if the shell is command.com or shellxquote is empty).

Change the prefixing spaces in menu items to appending spaces to make wildmenu users happier (when wim includes longest).